refactor: 빈 장바구니 상태 처리
